fix: move mapper to parser.go
