+++ internal/entitlement/httpdriver/entitlement.go
@@ -8,9 +8,6 @@ import (
 
 	"github.com/openmeterio/openmeter/api"
 	"github.com/openmeterio/openmeter/internal/entitlement"
-	booleanentitlement "github.com/openmeterio/openmeter/internal/entitlement/boolean"
-	meteredentitlement "github.com/openmeterio/openmeter/internal/entitlement/metered"
-	staticentitlement "github.com/openmeterio/openmeter/internal/entitlement/static"
 	"github.com/openmeterio/openmeter/internal/namespace/namespacedriver"
 	"github.com/openmeterio/openmeter/pkg/clock"
 	"github.com/openmeterio/openmeter/pkg/convert"
@@ -194,34 +191,6 @@ func (h *entitlementHandler) GetEntitlementValue() GetEntitlementValueHandler {
 			httptransport.WithErrorEncoder(getErrorEncoder()),
 		)...,
 	)
-}
-
-func MapEntitlementValueToAPI(entitlementValue entitlement.EntitlementValue) (api.EntitlementValue, error) {
-	switch ent := entitlementValue.(type) {
-	case *meteredentitlement.MeteredEntitlementValue:
-		return api.EntitlementValue{
-			HasAccess: convert.ToPointer(ent.HasAccess()),
-			Balance:   &ent.Balance,
-			Usage:     &ent.UsageInPeriod,
-			Overage:   &ent.Overage,
-		}, nil
-	case *staticentitlement.StaticEntitlementValue:
-		var config *string
-		if len(ent.Config) > 0 {
-			config = convert.ToPointer(string(ent.Config))
-		}
-
-		return api.EntitlementValue{
-			HasAccess: convert.ToPointer(ent.HasAccess()),
-			Config:    config,
-		}, nil
-	case *booleanentitlement.BooleanEntitlementValue:
-		return api.EntitlementValue{
-			HasAccess: convert.ToPointer(ent.HasAccess()),
-		}, nil
-	default:
-		return api.EntitlementValue{}, errors.New("unknown entitlement type")
-	}
 }
 
 type GetEntitlementsOfSubjectHandlerRequest = models.NamespacedID
