+++ internal/entitlement/httpdriver/parser.go
@@ -1,6 +1,7 @@
 package httpdriver
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/openmeterio/openmeter/api"
@@ -130,6 +131,34 @@ func (p parser) ToAPIGeneric(e *entitlement.Entitlement) (*api.Entitlement, erro
 		return res, nil
 	default:
 		return nil, fmt.Errorf("unsupported entitlement type: %s", e.EntitlementType)
+	}
+}
+
+func MapEntitlementValueToAPI(entitlementValue entitlement.EntitlementValue) (api.EntitlementValue, error) {
+	switch ent := entitlementValue.(type) {
+	case *meteredentitlement.MeteredEntitlementValue:
+		return api.EntitlementValue{
+			HasAccess: convert.ToPointer(ent.HasAccess()),
+			Balance:   &ent.Balance,
+			Usage:     &ent.UsageInPeriod,
+			Overage:   &ent.Overage,
+		}, nil
+	case *staticentitlement.StaticEntitlementValue:
+		var config *string
+		if len(ent.Config) > 0 {
+			config = convert.ToPointer(string(ent.Config))
+		}
+
+		return api.EntitlementValue{
+			HasAccess: convert.ToPointer(ent.HasAccess()),
+			Config:    config,
+		}, nil
+	case *booleanentitlement.BooleanEntitlementValue:
+		return api.EntitlementValue{
+			HasAccess: convert.ToPointer(ent.HasAccess()),
+		}, nil
+	default:
+		return api.EntitlementValue{}, errors.New("unknown entitlement type")
 	}
 }
